HiDPI awareness, variable scaling.
Doesn't currently account for the DPI so on HiDPI displays it will simply be smaller (or rather, on zoomed displays it won't zoom). The scale factor could be chosen for the current monitor's size so that it is e.g. as high as possible while fitting within the monitor's resolution.